requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
scikit-learn
opencv-python
imageio
Flask
